flutter pub upgrade --major-versions
This was less terrifying than it's been in the past, thanks to the
latest fix from @NamanShergill that upgrades our Dio caching. We
still have some outdated packages, but most of the major libs are
now in sync with the latest available.

With this upgrade, it feels like we can finally start working on
the actual UI! <3