Fixing Android icon (+ Gradle 7.3.0 upgrade)
It wasn't possible to build repeatedly on JDK 17 (though it worked sometimes, oddly enough) with Gradle 7.0.2 so Gradle is now 7.3.0.

I tried Gradle 7.4.0 and although it worked for one build, it refused
to build after I deleted some old 7.0.2 artifacts... not sure why.

The iOS and Android icons are now slightly different. Android uses
a slightly modified icon so it renders without touching the edges of
circles, squares, and "squircles" on the home screen.

I am _intentionally_ not using Android adaptive icons... they're
overly complicated and flat icons are preferable when we know we
want a white background behind the icon anyway.